requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0
mamba-ssm>=1.0.0
timm>=0.9.0
